+++ backend/src/main/java/com/codecool/trivia/model/request_schema/PointRequest.java
@@ -4,5 +4,5 @@ import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import com.fasterxml.jackson.annotation.JsonProperty;
 
 @JsonIgnoreProperties(ignoreUnknown = true)
-public record PointRequest(@JsonProperty String name, @JsonProperty double extraPoints) {
+public record PointRequest(@JsonProperty String name, @JsonProperty int extraPoints) {
 }
